edge/requirements.txt
pyserde==0.6.0
google-cloud-container==2.4.1
google-cloud-secret-manager==2.5.0